pyproject.toml
[tool.poetry]
name = "txt2pdf"
version = "0.1.0"
description = ""
authors = ["MaximPyanin <maximpyanin@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
celery = "^5.4.0"
fpdf2 = "^2.7.8"
dependency-injector = "^4.41.0"
gunicorn = "^22.0.0"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
